+++ authorization/grant_all/GrantAllGateway.go
@@ -1,9 +1,11 @@
 package grant_all
 
-import "github.com/attestify/go-kernel/authorization"
+import (
+	"github.com/attestify/go-kernel/authorization/access_control"
+)
 
 type GrantAllGateway interface {
-	Grant(control authorization.AccessControl)
+	Grant(control access_control.AccessControl)
 	Error() error
 	HasError() bool
 }
