+++ authorization/grant_all/GrantAll_test.go
@@ -2,7 +2,7 @@ package grant_all_test
 
 import (
 	"errors"
-	"github.com/attestify/go-kernel/authorization"
+	"github.com/attestify/go-kernel/authorization/access_control"
 	"github.com/attestify/go-kernel/authorization/grant_all"
 	"github.com/attestify/go-kernel/authorization/permission"
 	"github.com/attestify/go-kernel/error/internal_error"
@@ -187,7 +187,7 @@ func NewMockGrantAllGateway() MockGrantAllGateway {
 	return MockGrantAllGateway{}
 }
 
-func (gateway *MockGrantAllGateway) Grant(control authorization.AccessControl) {
+func (gateway *MockGrantAllGateway) Grant(control access_control.AccessControl) {
 	if gateway.grantAllGatewayInternalError {
 		gateway.mockError = internal_error.New("Error generated from GrantAll.Grant() invocation.")
 	}
