+++ authorization/grant_owner/GrantOwnerGateway.go
@@ -1,9 +1,11 @@
 package grant_owner
 
-import "github.com/attestify/go-kernel/authorization"
+import (
+	"github.com/attestify/go-kernel/authorization/access_control"
+)
 
 type GrantOwnerGateway interface {
-	Grant(accessControl authorization.AccessControl)
+	Grant(accessControl access_control.AccessControl)
 	HasError() bool
 	Error() error
 }
